# Cache routes received from `next_hop`